backend/requirements.txt
Django==3.2.16
python-dotenv==1.0.1
djangorestframework
djoser
django-extra-fields